Rename LoadoutGroupFiles to ItemContentsFileTree
