+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -622,6 +622,12 @@
         <Compile Update="Pages\NexusModsDataProvider.cs">
           <DependentUpon>ILibraryDataProvider.cs</DependentUpon>
         </Compile>
+        <Compile Update="Pages\ItemContentsFileTree\ItemContentsFileTreeViewModel.cs">
+          <DependentUpon>IItemContentsFileTreeViewModel.cs</DependentUpon>
+        </Compile>
+        <Compile Update="Pages\ItemContentsFileTree\ItemContentsFileTreeView.cs">
+          <DependentUpon>ItemContentsFileTreeView.axaml</DependentUpon>
+        </Compile>
     </ItemGroup>
 
     <ItemGroup>
