+++ src/NexusMods.App.UI/TypeFinder.cs
@@ -3,9 +3,9 @@ using NexusMods.App.UI.Pages.Changelog;
 using NexusMods.App.UI.Pages.Diagnostics;
 using NexusMods.App.UI.Pages.Diff.ApplyDiff;
 using NexusMods.App.UI.Pages.Downloads;
+using NexusMods.App.UI.Pages.ItemContentsFileTree;
 using NexusMods.App.UI.Pages.LibraryPage;
 using NexusMods.App.UI.Pages.LoadoutGrid;
-using NexusMods.App.UI.Pages.LoadoutGroupFiles;
 using NexusMods.App.UI.Pages.LoadoutPage;
 using NexusMods.App.UI.Pages.ModLibrary;
 using NexusMods.App.UI.Pages.MyGames;
@@ -36,7 +36,7 @@ internal class TypeFinder : ITypeFinder
         typeof(FileOriginsPageContext),
         typeof(TextEditorPageContext),
         typeof(MyLoadoutsPageContext),
-        typeof(LoadoutGroupFilesPageContext),
+        typeof(ItemContentsFileTreePageContext),
         typeof(LibraryPageContext),
         typeof(LoadoutPageContext),
